requirements.txt
paho-mqtt==1.6.1
pip==21.1.2
setuptools==57.0.0
socketcan==1.1.0
wheel==0.36.2
